refactor: 拆分 gateway 模块，将 C2C/群聊/缓存/流式发送抽离为独立文件
